feat: propagate procedure annotations/attrs to library/package
There are a number of useful things we can do with procedure annotations
(attributes), and much like type signatures, it is important to surface
this metadata in library metadata and in compiled packages.

This commit implements propagation of this information through assembly,
as well as all of the necessary serialization infrastructure for the
types involved.