fix: remove serialization of Span via serde_spanned
The trick used by serde_spanned is not supported by all deserializers,
and in unsupported serializers (such as serde_json), deserialization
will fail because the shape of the serialized value differs from the one
expected at deserialization time.

We were using this to provide more informative source spans when
deserializing our types from TOML, so this will produce worse errors in
the near term for that purpose - I will revisit improving that after
doing some research on how to implement this in a more general way.