ci: make contribution quality job not trigger on local branches
PR #2333 modified the author_association trigger to make the job fire
on the "contributor" value, which is applied overly broadly. As a way
to protect maintainers of this repo from contribution quality nagging,
this skips the job on a PR opened from the repo itself, as opposed to
a fork.

See https://docs.github.com/en/rest/pulls/pulls?apiVersion=2022-11-28 for details.